Add files via upload
This is a semester project for microcontroller and embedded system in which we have used arduino mega, encoders for moters, ultrasonic sensor,  Infrared sensor, voltage, current semester, L298N motor driver, Esp 32, and a wifi and an sd card module. The code is custom Atmel (software) based and the language is C. The code is in the report.